[blackboard] switch-in blackboard v2
