+++ .settings/org.eclipse.core.resources.prefs
@@ -8,4 +8,4 @@ encoding//doc/examples/selector.py=utf-8
 encoding//doc/examples/sequence.py=utf-8
 encoding//doc/examples/skeleton_behaviour.py=utf-8
 encoding//doc/examples/skeleton_tree.py=utf-8
-encoding//py_trees/blackboard2.py=utf-8
+encoding//py_trees/blackboard.py=utf-8
