+++ py_trees/blackboard.py
@@ -1,4 +1,5 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 #
 # License: BSD
 #   https://raw.githubusercontent.com/splintered-reality/py_trees/devel/LICENSE
@@ -8,26 +9,36 @@
 ##############################################################################
 
 """
-Blackboards are not a necessary component, but are a fairly standard feature
-in most behaviour tree implementations. See, for example, the `design notes`_
+Blackboards are not a necessary component of behaviour tree implementations,
+but are nonetheless, a fairly common mechanism for for sharing data between
+behaviours in the tree. See, for example, the `design notes`_
 for blackboards in Unreal Engine.
 
 .. image:: images/blackboard.jpg
    :width: 300px
    :align: center
 
-Implementations however, tend to vary quite a bit depending on the needs of
-the framework using them. Some of the usual considerations include scope
-and sharing of blackboards across multiple tree instances.
+Implementations vary widely depending on the needs of
+the framework using them. The simplest implementations take the
+form of a key-value store with global access, while more
+rigorous implementations scope access and form a secondary
+graph overlaying the tree graph connecting data ports between behaviours.
 
-For this package, we've decided to keep blackboards extremely simple to fit
-with the same 'rapid development for small scale systems' principles
-that this library is designed for.
+The implementation here strives to remain simple to use
+(so 'rapid development' does not become just 'development'), yet
+sufficiently featured so that the magic behind the scenes (i.e. the
+data sharing on the blackboard) is exposed and helpful in debugging
+tree applications.
 
-* No sharing between tree instances
-* No locking for reading/writing
-* Global scope, i.e. any behaviour can access any variable
-* No external communications (e.g. to a database)
+To be more concrete, the following is a list of features that this
+implementation either embraces or does not.
+
+* [+] Centralised key-value store
+* [+] Client based usage with declaration of read/write intentions at construction
+* [+] Activity stream that tracks read/write operations by behaviours
+* [-] Sharing between tree instances
+* [-] Locking for reading/writing
+* [-] Strict variable initialisation policies
 
 .. include:: weblinks.rst
 """
@@ -36,7 +47,11 @@ that this library is designed for.
 # Imports
 ##############################################################################
 
+import enum
+import re
 import operator
+import typing
+import uuid
 
 from . import behaviour
 from . import behaviours
@@ -48,140 +63,501 @@ from . import console
 ##############################################################################
 
 
+class KeyMetaData(object):
+
+    def __init__(self):
+        self.read = set()
+        self.write = set()
+
+
+class ActivityType(enum.Enum):
+    """An enumerator representing the operation on a blackboard variable"""
+
+    READ = "READ"
+    """Read from the blackboard"""
+    INITIALISED = "INITIALISED"
+    """Initialised value on the blackboard"""
+    WRITE = "WRITE"
+    """Wrote to the blackboard."""
+    READ_FAILED = "READ_FAILED"
+    """Tried to read a key that does not yet exist on the blackboard."""
+    READ_DENIED = "READ_DENIED"
+    """Client was denied access to read a key."""
+    WRITE_DENIED = "WRITE_DENIED"
+    """Client was denied access to write a key."""
+    NO_OVERWRITE = "NO_OVERWRITE"
+    """Variable already exists and a no-overwrite request was respected."""
+    UNSET = "UNSET"
+    """Key was removed from the blackboard"""
+
+
+class ActivityItem(object):
+    """
+    Recorded data pertaining to activity on the blackboard.
+
+    Args:
+        key: name of the variable on the blackboard
+        client_name: convenient name of the client performing the operation
+        client_id: unique id of the client performing the operation
+        activity_type: type of activity
+        previous_value: of the given key (None if this field is not relevant)
+        current_value: current value for the given key (None if this field is not relevant)
+    """
+    def __init__(
+            self,
+            key,
+            client_name: str,
+            client_id: uuid.UUID,
+            activity_type: ActivityType,
+            previous_value: typing.Any=None,
+            current_value: typing.Any=None):
+        # TODO validity checks for values passed/not passed on the
+        # respective activity types. Note: consider using an enum
+        # for 'no value here' since None is a perfectly acceptable
+        # value for a key
+        self.key = key
+        self.client_name = client_name
+        self.client_id = client_id
+        self.activity_type = activity_type
+        self.previous_value = previous_value
+        self.current_value = current_value
+
+
+class ActivityStream(object):
+    """
+    Storage container with convenience methods for manipulating the stored
+    activity stream.
+
+    Attributes:
+        data (typing.List[ActivityItem]: list of activity items, earliest first
+        maximum_size (int): pop items if this size is exceeded
+    """
+
+    def __init__(self, maximum_size: int=500):
+        """
+        Initialise the stream with a maximum storage limit.
+
+        Args:
+            maximum_size: pop items from the stream if this size is exceeded
+        """
+        self.data = []
+        self.maximum_size = maximum_size
+
+    def push(self, activity_item: ActivityItem):
+        """
+        Push the next activity item to the stream.
+
+        Args:
+            activity_item: new item to append to the stream
+        """
+        if len(self.data) > self.maximum_size:
+            self.data.pop()
+        self.data.append(activity_item)
+
+    def clear(self):
+        """
+        Delete all activities from the stream.
+        """
+        self.data = []
+
+
 class Blackboard(object):
     """
-    `Borg`_ style key-value store for sharing amongst behaviours.
-
-    .. _Borg: http://code.activestate.com/recipes/66531-singleton-we-dont-need-no-stinkin-singleton-the-bo/
+    Key-value store for sharing data between behaviours.
 
     Examples:
-        You can instantiate the blackboard from anywhere in your program. Even
-        disconnected calls will get access to the same data store. For example:
+        You can instantiate a blackboard client anywhere in your program.
+        Even disconnected instantiations will discover the single global key-value store.
+        For example:
 
         .. code-block:: python
 
             def check_foo():
-                blackboard = Blackboard()
+                blackboard = Blackboard(name="Reader", read={"foo"))
                 assert(blackboard.foo, "bar")
 
             if __name__ == '__main__':
-                blackboard = Blackboard()
+                blackboard = Blackboard(name="Writer", write={"foo"))
                 blackboard.foo = "bar"
                 check_foo()
 
-        If the key value you are interested in is only known at runtime, then
-        you can set/get from the blackboard without the convenient variable style
-        access:
+        To respect an already initialised key on the blackboard:
 
         .. code-block:: python
 
-            blackboard = Blackboard()
-            result = blackboard.set("foo", "bar")
-            foo = blackboard.get("foo")
+            if __name__ == '__main__':
+                blackboard = Blackboard(name="Writer", read={"foo"))
+                result = blackboard.set("foo", "bar", overwrite=False)
 
-        The blackboard can also be converted and printed (with highlighting)
-        as a string. This is useful for logging and debugging.
+        Log and display the activity stream:
 
         .. code-block:: python
 
-            print(Blackboard())
+            # all key-value pairs
+            Blackboard.enable_activity_stream(maximum_size=100)
+            blackboard = Blackboard(name="Writer", write={"foo"))
+            blackboard.foo = "bar"
+            # view the activity stream
+            print(py_trees.display.unicode_blackboard_activity_stream())
+            # output
+            Activity Stream
+                foo       : INITIALISED  | Set Foo   | → bar
+            # clear the stream (useful to do between tree ticks)
+            Blackboard.activity_stream.clear()
 
+        For introspection, use the methods in the display module:
+
+        .. code-block:: python
+
+            # all key-value pairs
+            print(py_trees.display.unicode_blackboard())
+            # various filtered views
+            print(py_trees.display.unicode_blackboard(key_filter={"foo"}))
+            print(py_trees.display.unicode_blackboard(regex_filter="dud*"))
+            print(py_trees.display.unicode_blackboard(client_filter={writer.id, set_foo.id}))
+            # list the clients associated with each key
+            print(py_trees.display.unicode_blackboard(display_only_key_metadata=True)
+            # a dot graph representation of tree and blackboard together
+            py_trees.display.render_dot_tree(root, root, with_blackboard_variables=True)
 
     .. warning::
 
        Be careful of key collisions. This implementation leaves this management up to the user.
 
-    .. seealso:: The :ref:`py-trees-demo-blackboard-program` program demos use of the blackboard along with a couple of the blackboard behaviours.
+    .. seealso::
+       * :ref:`py-trees-demo-blackboard <py-trees-demo-blackboard-program>`
+
+    Attributes:
+        <class>.storage (typing.Dict[str, typing.Any]): key-value storage
+        <class>.read (typing.Dict[str, typing.List[uuid.UUID]]): key-client read access map
+        <class>.write (typing.Dict[str, typing.List[uuid.UUID]]): key-client write access map
+        <class>.activity_stream (ActivityStream): the activity stream (None if not enabled)
+        <instance>.name (str): client's convenient, but not necessarily unique identifier
+        <instance>.unique_identifier (uuid.UUID): client's unique identifier
+        <instance>.read (typing.List[str]): keys this client has permission to read
+        <instance>.write (typing.List[str]): keys this client has permission to write
+
+    Args:
+        name: the non-unique, but convenient identifier (stringifies the uuid if None) for the client
+        unique_identifier: client's unique identifier (auto-generates if None)
+        read: list of keys this client has permission to read
+        write: list of keys this client has permission to write
+
+    Raises:
+        TypeError: if the provided name/unique identifier is not of type str/uuid.UUID
+        ValueError: if the unique identifier has already been registered
+
     """
-    # Dunder style to avoid collisions
-    __shared_state = {}
+    storage = {}  # Dict[str, Any] / key-value storage
+    metadata = {}  # Dict[ str, KeyMetaData ] / key-metadata information
+    clients = {}   # Dict[ uuid.UUID, Blackboard] / id-client information
+    activity_stream = None
 
-    @staticmethod
-    def clear():
-        """
-        Erase the blackboard contents. Typically this is used only when you
-        have repeated runs of different tree instances, as often happens in testing.
-        """
-        Blackboard.__shared_state.clear()
+    def __init__(
+            self, *,
+            name: str=None,
+            unique_identifier: uuid.UUID=None,
+            read: typing.List[str]=[],
+            write: typing.List[str]=[]):
+        # print("__init__")
+        if unique_identifier is None:
+            unique_identifier = uuid.uuid4()
+        if type(unique_identifier) != uuid.UUID:
+            raise TypeError("provided unique identifier is not of type uuid.UUID")
+        if name is None or not name:
+            name = str(unique_identifier)
+        if not isinstance(name, str):
+            raise TypeError("provided name is not of type str [{}]".format(type(name)))
+        super().__setattr__("unique_identifier", unique_identifier)
+        super().__setattr__("name", name)
+        super().__setattr__("read", read)
+        for key in read:
+            Blackboard.metadata.setdefault(key, KeyMetaData())
+            Blackboard.metadata[key].read.add(
+                super().__getattribute__("unique_identifier")
+            )
+        super().__setattr__("write", write)
+        for key in write:
+            Blackboard.metadata.setdefault(key, KeyMetaData())
+            Blackboard.metadata[key].write.add(
+                super().__getattribute__("unique_identifier")
+            )
+        Blackboard.clients[
+            super().__getattribute__("unique_identifier")
+        ] = self
 
-    def __init__(self):
-        self.__dict__ = self.__shared_state
-
-    def set(self, name, value, overwrite=True):
+    def __setattr__(self, name, value):
         """
-        For when you only have strings to identify and access the blackboard variables, this
-        provides a convenient setter.
+        Convenience attribute style referencing with checking against
+        permissions.
+
+        Raises:
+            AttributeError: if the client does not have write access to the variable
+        """
+        # print("__setattr__ [{}][{}]".format(name, value))
+        if name not in super().__getattribute__("write"):
+            if Blackboard.activity_stream is not None:
+                Blackboard.activity_stream.push(
+                    self._generate_activity_item(name, ActivityType.WRITE_DENIED)
+                )
+            raise AttributeError("client '{}' does not have write access to '{}'".format(self.name, name))
+        if Blackboard.activity_stream is not None:
+            if name in Blackboard.storage.keys():
+                Blackboard.activity_stream.push(
+                    self._generate_activity_item(
+                        key=name,
+                        activity_type=ActivityType.WRITE,
+                        previous_value=Blackboard.storage[name],
+                        current_value=value
+                    )
+                )
+            else:
+                Blackboard.activity_stream.push(
+                    self._generate_activity_item(
+                        key=name,
+                        activity_type=ActivityType.INITIALISED,
+                        current_value=value
+                    )
+                )
+        Blackboard.storage[name] = value
+
+    def __getattr__(self, name):
+        """
+        Convenience attribute style referencing with checking against
+        permissions.
+
+        Raises:
+            AttributeError: if the client does not have read access to the variable
+            ValueError: if the variable does not yet exist on the blackboard
+        """
+        # print("__getattr__ [{}]".format(name))
+        try:
+            if name not in super().__getattribute__("read"):
+                if Blackboard.activity_stream is not None:
+                    Blackboard.activity_stream.push(
+                        self._generate_activity_item(name, ActivityType.READ_DENIED)
+                    )
+                raise AttributeError("client '{}' does not have read access to '{}'".format(self.name, name))
+            if Blackboard.activity_stream is not None:
+                Blackboard.activity_stream.push(
+                    self._generate_activity_item(
+                        key=name,
+                        activity_type=ActivityType.READ,
+                        current_value=Blackboard.storage[name],
+                    )
+                )
+            return Blackboard.storage[name]
+        except KeyError as e:
+            if Blackboard.activity_stream is not None:
+                Blackboard.activity_stream.push(
+                    self._generate_activity_item(name, ActivityType.READ_FAILED)
+                )
+            raise ValueError("variable '{}' does not yet exist on the blackboard".format(name)) from e
+
+    def set(self, name: str, value: typing.Any, overwrite: bool=True) -> bool:
+        """
+        Set, conditionally depending on whether the variable already exists or otherwise.
+
+        This is most useful when initialising variables and multiple elements
+        seek to do so. A good policy to adopt for your applications in these situations is
+        a first come, first served policy. Ensure global configuration has the first
+        opportunity followed by higher priority behaviours in the tree and so forth.
+        Lower priority behaviours would use this to respect the pre-configured
+        setting and at most, just validate that it is acceptable to the functionality
+        of it's own behaviour.
 
         Args:
-            name (:obj:`str`): name of the variable to set
-            value (:obj:`any`): any variable type
-            overwrite(:obj:`bool`): whether to abort if the value is already present
+            name: name of the variable to set
+            value: value of the variable to set
+            overwrite: do not set if the variable already exists on the blackboard
 
         Returns:
-            :obj:`bool`: always True unless overwrite was set to False and a variable already exists
+            success or failure (overwrite is False and variable already set)
+
+        Raises:
+            AttributeError: if the client does not have write access to the variable
         """
+        if name not in super().__getattribute__("write"):
+            if Blackboard.activity_stream is not None:
+                Blackboard.activity_stream.push(
+                    self._generate_activity_item(name, ActivityType.WRITE_DENIED)
+                )
+            raise AttributeError("client does not have write access to '{}'".format(name))
         if not overwrite:
-            try:
-                getattr(self, name)
+            if name in Blackboard.storage:
+                if Blackboard.activity_stream is not None:
+                    Blackboard.activity_stream.push(
+                        self._generate_activity_item(
+                            key=name,
+                            activity_type=ActivityType.NO_OVERWRITE,
+                            current_value=Blackboard.storage[name])
+                    )
                 return False
-            except AttributeError:
-                pass
         setattr(self, name, value)
         return True
 
-    def get(self, name):
+    def get(self, name: str) -> typing.Any:
         """
-        For when you only have strings to identify and access the blackboard variables,
-        this provides a convenient accessor.
+        Method based accessor to the blackboard variables (as opposed to simply using
+        '.<name>').
 
         Args:
-            name (:obj:`str`): name of the variable to get
-        """
-        try:
-            return getattr(self, name)
-        except AttributeError:
-            return None
+            name: name of the variable to get
 
-    def unset(self, name):
+        Raises:
+            AttributeError: if the client does not have read access to the variable
+            ValueError: if the variable does not yet exist on the blackboard
         """
-        For when you need to unset a blackboard variable, this provides a convenient helper method.
-        This is particularly useful for unit testing behaviours.
+        getattr(self, name)
+
+    def unset(self, name: str):
+        """
+        For when you need to completely remove a blackboard variable,
+        this provides a convenient helper method. This is particularly
+        useful when you wish to 'erase' the blackboard between unit test
+        methods.
 
         Args:
-            name (:obj:`str`): name of the variable to unset
+            name: name of the variable to clear
+
+        Raises:
+            KeyError: if the variable does not yet exist
         """
-        try:
-            delattr(self, name)
-        except AttributeError:
-            pass
+        if Blackboard.activity_stream is not None:
+            Blackboard.activity_stream.push(
+                self._generate_activity_item(name, ActivityType.UNSET)
+            )
+        del Blackboard.storage[name]
 
     def __str__(self):
-        """
-        Express the blackboard contents as a string. Useful for debugging.
+        indent = "  "
+        s = console.green + type(self).__name__ + console.reset + "\n"
+        s += console.white + indent + "Client Data" + console.reset + "\n"
+        keys = ["name", "unique_identifier", "read", "write"]
+        s += self._stringify_key_value_pairs(keys, self.__dict__, 2 * indent)
+        s += console.white + indent + "Variables" + console.reset + "\n"
+        keys = list(dict.fromkeys(self.read + self.write))  # unique list, https://www.peterbe.com/plog/fastest-way-to-uniquify-a-list-in-python-3.6
+        s += self._stringify_key_value_pairs(keys, Blackboard.storage, 2 * indent)
+        return s
 
-        Returns:
-            :obj:`str`: blackboard contents
-        """
-        s = console.green + type(self).__name__ + "\n" + console.reset
+    def _generate_activity_item(self, key, activity_type, previous_value=None, current_value=None):
+        return ActivityItem(
+            key=key,
+            client_name=super().__getattribute__("name"),
+            client_id=super().__getattribute__("unique_identifier"),
+            activity_type=activity_type,
+            previous_value=previous_value,
+            current_value=current_value
+        )
+
+    def _stringify_key_value_pairs(self, keys, key_value_dict, indent):
+        s = ""
         max_length = 0
-        for k in self.__dict__.keys():
-            max_length = len(k) if len(k) > max_length else max_length
-        keys = sorted(self.__dict__)
         for key in keys:
-            value = self.__dict__[key]
-            if value is None:
-                value_string = "-"
-                s += console.cyan + "  " + '{0: <{1}}'.format(key, max_length + 1) + console.reset + ": " + console.yellow + "{0}\n".format(value_string) + console.reset
-            else:
+            max_length = len(key) if len(key) > max_length else max_length
+        for key in keys:
+            try:
+                value = key_value_dict[key]
                 lines = ('{0}'.format(value)).split('\n')
                 if len(lines) > 1:
-                    s += console.cyan + "  " + '{0: <{1}}'.format(key, max_length + 1) + console.reset + ":\n"
+                    s += console.cyan + indent + '{0: <{1}}'.format(key, max_length + 1) + console.reset + ":\n"
                     for line in lines:
-                        s += console.yellow + "    {0}\n".format(line) + console.reset
+                        s += console.yellow + indent + "  {0}\n".format(line) + console.reset
                 else:
-                    s += console.cyan + "  " + '{0: <{1}}'.format(key, max_length + 1) + console.reset + ": " + console.yellow + '{0}\n'.format(value) + console.reset
+                    s += console.cyan + indent + '{0: <{1}}'.format(key, max_length + 1) + console.reset + ": " + console.yellow + '{0}\n'.format(value) + console.reset
+            except KeyError:
+                s += console.cyan + indent + '{0: <{1}}'.format(key, max_length + 1) + console.reset + ": " + console.yellow + "-\n" + console.reset
         s += console.reset
         return s
+
+    def unregister(self, clear=True):
+        """
+        Unregister this blackboard and if requested, clear key-value pairs if this
+        client is the last user of those variables.
+        """
+        for key in self.read:
+            Blackboard.metadata[key].read.remove(super().__getattribute__("unique_identifier"))
+        for key in self.write:
+            Blackboard.metadata[key].write.remove(super().__getattribute__("unique_identifier"))
+        if clear:
+            for key in (set(self.read) | set(self.write)):
+                if not (set(Blackboard.metadata[key].read) | set(Blackboard.metadata[key].write)):
+                    Blackboard.storage.pop(key, None)
+
+    @staticmethod
+    def keys() -> typing.Set[str]:
+        """
+        Get the set of blackboard keys.
+
+        Returns:
+            the complete set of keys registered by clients
+        """
+        # return registered keys, those on the blackboard are not
+        # necessarily written to yet
+        return Blackboard.metadata.keys()
+
+    @staticmethod
+    def keys_filtered_by_regex(regex: str) -> typing.Set[str]:
+        """
+        Get the set of blackboard keys filtered by regex.
+
+        Args:
+            regex: a python regex string
+
+        Returns:
+            subset of keys that have been registered and match the pattern
+        """
+        pattern = re.compile(regex)
+        return [key for key in Blackboard.metadata.keys() if pattern.search(key) is not None]
+
+    @staticmethod
+    def keys_filtered_by_clients(client_ids: typing.Union[typing.List[str], typing.Set[str]]) -> typing.Set[str]:
+        """
+        Get the set of blackboard keys filtered by client ids.
+
+        Args:
+            client_ids: set of client uuid's.
+
+        Returns:
+            subset of keys that have been registered by the specified clients
+        """
+        # convenience for users
+        if type(client_ids) == list:
+            client_ids = set(client_ids)
+        keys = set()
+        for key in Blackboard.metadata.keys():
+            # for sets, | is union, & is intersection
+            key_clients = set(Blackboard.metadata[key].read) | set(Blackboard.metadata[key].write)
+            if key_clients & client_ids:
+                keys.add(key)
+        return keys
+
+    @staticmethod
+    def enable_activity_stream(maximum_size: int=500):
+        """
+        Enable logging of activities on the blackboard.
+
+        Args:
+            maximum_size: pop items from the stream if this size is exceeded
+
+        Raises:
+            RuntimeError if the activity stream is already enabled
+        """
+        if Blackboard.activity_stream is None:
+            Blackboard.activity_stream = ActivityStream(maximum_size)
+        else:
+            RuntimeError("activity stream is already enabled for this blackboard")
+
+    @staticmethod
+    def disable_activity_stream():
+        """
+        Disable logging of activities on the blackboard
+        """
+        Blackboard.activity_stream = None
+
+##############################################################################
+# Blackboard Behaviours
+##############################################################################
 
 
 class ClearBlackboardVariable(behaviours.Success):
@@ -198,12 +574,15 @@ class ClearBlackboardVariable(behaviours.Success):
                  ):
         super(ClearBlackboardVariable, self).__init__(name)
         self.variable_name = variable_name
+        self.blackboard = Blackboard(
+            name=self.name,
+            write={self.variable_name}
+        )
 
     def initialise(self):
         """
         Delete the variable from the blackboard.
         """
-        self.blackboard = Blackboard()
         self.blackboard.unset(self.variable_name)
 
 
@@ -234,9 +613,13 @@ class SetBlackboardVariable(behaviours.Success):
         super(SetBlackboardVariable, self).__init__(name)
         self.variable_name = variable_name
         self.variable_value = variable_value
+        self.blackboard = Blackboard(
+            name=self.name,
+            unique_identifier=self.id,
+            write={self.variable_name}
+        )
 
     def initialise(self):
-        self.blackboard = Blackboard()
         self.blackboard.set(self.variable_name, self.variable_value, overwrite=True)
 
 
@@ -278,8 +661,12 @@ class CheckBlackboardVariable(behaviour.Behaviour):
             subgoal. Use the :data:`~py_trees.common.ClearingPolicy.NEVER` flag to do this.
         """
         super(CheckBlackboardVariable, self).__init__(name)
-        self.blackboard = Blackboard()
         self.variable_name = variable_name
+        self.blackboard = Blackboard(
+            name=self.name,
+            unique_identifier=self.id,
+            read={self.variable_name}
+        )
         self.expected_value = expected_value
         self.comparison_operator = comparison_operator
         self.matching_result = None
@@ -385,8 +772,12 @@ class WaitForBlackboardVariable(behaviour.Behaviour):
                  clearing_policy=common.ClearingPolicy.ON_INITIALISE
                  ):
         super(WaitForBlackboardVariable, self).__init__(name)
-        self.blackboard = Blackboard()
         self.variable_name = variable_name
+        self.blackboard = Blackboard(
+            name=self.name,
+            unique_identifier=self.id,
+            read={self.variable_name}
+        )
         self.expected_value = expected_value
         self.comparison_operator = comparison_operator
         self.clearing_policy = clearing_policy
@@ -447,3 +838,61 @@ class WaitForBlackboardVariable(behaviour.Behaviour):
         self.logger.debug("%s.terminate(%s)" % (self.__class__.__name__, "%s->%s" % (self.status, new_status) if self.status != new_status else "%s" % new_status))
         if new_status == common.Status.INVALID:
             self.matching_result = None
+
+
+##############################################################################
+# Main
+##############################################################################
+
+
+def main():
+    bb2 = Blackboard(
+        name="Bob's Board",
+        unique_identifier=uuid.uuid4(),
+        read=['foo', 'bar'],
+        write=['foo', 'dude'],
+    )
+    bb2.foo = "more bar"
+    bb2.dude = "bob"
+    unused_myfoo = bb2.foo
+    bb2.introspect_blackboard()
+    print("-----------------")
+    print("{}".format(str(bb2)))
+    print("-----------------")
+    print(console.green + "Exceptions\n" + console.reset)
+    try:
+        print(console.green + "  get(<key>): doesn't exist on the blackboard" + console.reset)
+        print("bar: {}".format(bb2.get('bar')))
+    except Exception as e:
+        print("    {}: {}".format(type(e), str(e)))
+    try:
+        print(console.green + "  get(<key>): doesn't have read access..." + console.reset)
+        print("dude: {}".format(bb2.get('dude')))
+    except Exception as e:
+        print("    {}: {}".format(type(e), str(e)))
+    try:
+        print(console.green + "  .<key>: doesn't exist on the blackboard" + console.reset)
+        print("bar: {}".format(bb2.bar))
+    except Exception as e:
+        print("    {}: {}".format(type(e), str(e)))
+    try:
+        print(console.green + "  .<key>: doesn't have read access..." + console.reset)
+        print("dude: {}".format(bb2.dude))
+    except Exception as e:
+        print("    {}: {}".format(type(e), str(e)))
+    # set
+    try:
+        print(console.green + "  set(<key>): doesn't have write permissions" + console.reset)
+        bb2.set('foobar', 3)
+    except Exception as e:
+        print("    {}: {}".format(type(e), str(e)))
+    try:
+        print(console.green + "  set(<key>): could not overwrite existing variable" + console.reset)
+        bb2.set('foo', 3, overwrite=False)
+    except Exception as e:
+        print("    {}: {}".format(type(e), str(e)))
+    try:
+        print(console.green + "  .<key>=...: doesn't have write permissions" + console.reset)
+        bb2.foobar = 3
+    except Exception as e:
+        print("    {}: {}".format(type(e), str(e)))
