+++ tests/test_eternal_guard.py
@@ -101,8 +101,11 @@ def test_eternal_guard_idiom():
 
 
 def test_eternal_guard_unique_names():
-    blackboard = py_trees.blackboard.Blackboard()
-    blackboard.clear()
+    blackboard = py_trees.blackboard.Blackboard(
+        read={"eternal_guard_condition_1",
+              "eternal_guard_condition_2"},
+        write={"eternal_guard_condition_1"}
+    )
     message = "Ha, stole it"
     blackboard.eternal_guard_condition_1 = message
     root = py_trees.composites.Selector(name="Root")
@@ -111,6 +114,7 @@ def test_eternal_guard_unique_names():
     # tick once, get variables on the blackboard
     py_trees.tests.tick_tree(root, 1, 1, print_snapshot=True)
     assert(blackboard.get("eternal_guard_condition_1") == message)  # wasn't overwritten
+    print(py_trees.display.unicode_blackboard())
     assert(blackboard.get("eternal_guard_condition_2") is None)
     # can't assert on the uuid variables though - they could be anything
